requirements.txt
gradio==5.12.0
numpy==2.0.0
torch==2.5.1
torchaudio==2.5.1
transformers==4.48.0
llama_cpp_python==0.3.7
pydub==0.25.1
accelerate==1.2.1
pygame==2.6.1
violet-simulator==0.2.4
matplotlib==3.10.0
nltk==3.9.1
rouge-score==0.1.2
evaluate==0.4.3
protobuf==5.29.3
sentencepiece==0.2.0
librosa==0.10.2.post1
py_trees==2.3.0
pyttsx3==2.98